try multiple figures, change short to int, add check for t1 and t2 in the sphere and add a comment

+++ include/intersection.h
@@ -15,7 +15,7 @@
 # include "types.h"
 # include "vec3_operations.h"
 
-short sphere_intersection(t_ray ray, t_scene_element *elem, int bounce);
-short plane_intersection(t_ray ray, t_scene_element *elem, int bounce);
-short cylinder_intersection(t_ray ray, t_scene_element *elem, int bounce);
+int sphere_intersection(t_ray ray, t_scene_element *elem, int bounce);
+int plane_intersection(t_ray ray, t_scene_element *elem, int bounce);
+int cylinder_intersection(t_ray ray, t_scene_element *elem, int bounce);
 #endif //INTERSECTION_H
